Conversion of the Federated Learning Server to Rx.
This in-place convertes the future based version of the server into a javarx based on. It temporarily deactivates boq conformance tests while we discuss with the boq team.

CL navigation:

- The package com.google.intelligence.micore.common.rx contains a couple of utilities to work with Rx. Specifically, (a) converting an RpcFutureStream into a pair of Observable/Observer (b) a declarative approach to define state machines for creating observers. It turns out that Rx is rather neutral and doesn't really has much support for creating observers (but lots for consuming observables) (c) dealing with conversion of futures.

- The package com.google.knowledge.cerebra.quirk.federated contains the new server code. There are three new classes which together implement the original, monolithic PopulationServlet:

  (1) The ClientActor which is the actor which handles all incoming requests from the stream and produces outputs.

  (2) The ParticipantSelector (and subclasses) which implements the strategy for selecting clients for a training phase.

  (3) The PhaseActor which is the actor which coordinates a training phase.

A kind of orthogonal change to the Rx migration is removing the singleton  lifetime of the PopulationServlet. Its replacement, TrainingPhase, lives for just one phase which makes the config update logic simpler. If we actually need to persist state across phases, we should do so by other means.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=131370687