Adding ignore clauses for Documentation and System Types
Documentation will be ignored because the changes are incredibly verbose and
generally of no interest to the users. Doc changes can never be breaking
changes (from the perspective pushAdvisor is interested in), and is rarely or
never of interest to customers

System Types are really an internal detail since they are non customer defined
types. Changes here should not be of interest to customers, similar to
derived_data.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=135984302