Fix Proto type declaration generation logic
Previously, the heuristic used to generate the type declaration for a proto field was sometimes incorrect. This change adds a check so that, if the heuristic fails, the tool uses the fully-qualified, dot-prefixed name for the type, which is always safe.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=136035382