Undo the removal of empty messages from baseline tests.
We are using the same baseline files to test the opensource and non opensouced api compiler. Therefore stripping the empty message causes opensouce baseline comparison to fail in cases where all the inner fields of the message are GOOGLE_INTERNAL. When creating the baseline in the non opensourced world, the containing message will not get stripped but will have all its fields MOE stripped out. But when the same test is executed in Opensouce, since the fields of the outer message are GOOGLE_INTERNAL, the message's content become empty and it is not present in the output when comparing the baseline files).

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=146314544