Inception api compiler validation would fail when audience is set inside auth provider.
This was happening because we are copying audience from auth provider to auth requirement (as part of avoiding changing the implementation of ESP). This works fine inside the api compiler. However, when the service configuration is passed to inception (or submit config source when creates the service config and then does validation on it), the audience fields now is present on both provider and requirement, causing this validation error.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=146617580