Message field inlining
Introduces a new inlining Instruction `(== inline_message ==)` that can be attached to proto messages. When a message is inlined, DocGen will not produce a documentation page for that message. Instead, wherever the message type is referenced, the messages contents are inlined at the site of the reference; no links to the type are made. Recursive messages are not inline-able to prevent infinite inlining.

Inlining is implemented in the following places:
- Response message
- Path param
- Query param
- Request body

Staged examples of inlining the "Book" type from the library example:

with inlining configured
  https://devsite.googleplex.com/library-example/rest/v1/shelves.books/create
  https://screenshot.googleplex.com/nkJWuCw7JKE

without inlining
  https://devsite.googleplex.com/library-example2/library-example/rest/v1/shelves.books/create
  https://screenshot.googleplex.com/QfrbKmcw0pk

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=151828421