Moves CEL compiler out of contracts for independent usage.
Tests are still via the contract compiler. which shall be fixed in a subsequent CL.

This prepares to extend the CEL compiler, which is based on OP tools framework proto model and snippet template system, to other languages (specifically C++). Contracts require a compiler for best user experience, so this is orthogonal to other work on evaluators. However, it's better to isolate this compiler from contracts itself, so it may be useful also in other applications.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=153546149