Allow non-unique collection short names across versions
Previously, all collection short names were required to be unique. This is not possible when collections are versioned, as there may be a "copy" of any given collection in any given version, and it is not a conflict to have both "v1.foos" and "v2.foos" share the short name "foos". (The requirement that collection short names are unique within a particular version remains.)

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=165471261